exception handling

+++ src/main/java/org/kalipo/web/rest/CommentResource.java
@@ -1,6 +1,8 @@
 package org.kalipo.web.rest;
 
 import com.codahale.metrics.annotation.Timed;
+import com.wordnik.swagger.annotations.*;
+import org.apache.commons.lang3.StringUtils;
 import org.kalipo.domain.Comment;
 import org.kalipo.repository.CommentRepository;
 import org.kalipo.security.SecurityUtils;
@@ -17,6 +19,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.inject.Inject;
 import javax.validation.Valid;
 import javax.ws.rs.QueryParam;
+import java.security.InvalidParameterException;
 import java.util.List;
 import java.util.Optional;
 
@@ -25,6 +28,7 @@ import java.util.Optional;
  */
 @RestController
 @RequestMapping("/app")
+@Api(value = "/app", description = "Operations about comments")
 public class CommentResource {
 
     private final Logger log = LoggerFactory.getLogger(CommentResource.class);
@@ -33,33 +37,41 @@ public class CommentResource {
     private CommentRepository commentRepository;
 
     /**
-     * PUT  /rest/comments -> Create a new comment.
+     * POST  /rest/comments -> Create a new comment.
      */
     @RequestMapping(value = "/rest/comments",
-            method = RequestMethod.PUT,
+            method = RequestMethod.POST,
             produces = MediaType.APPLICATION_JSON_VALUE)
     @Timed
-    public void create(@RequestBody CommentDTO commentDTO) {
+    @ResponseStatus(HttpStatus.CREATED)
+    @ApiOperation(value = "Create a new comment")
+    public void create(@Valid @RequestBody CommentDTO commentDTO) {
         log.debug("REST request to save Comment : {}", commentDTO);
 
         Comment comment = CommentDTO.convert(commentDTO);
-        comment.setAuthorId("d");
+        comment.setAuthorId("d"); // todo SecurityUtils.getCurrentLogin() is null during tests
         comment.setStatus(Comment.Status.APPROVED);
         commentRepository.save(comment);
     }
 
     /**
-     * POST  /rest/comments -> Update existing comment.
+     * PUT  /rest/comments -> Update existing comment.
      */
     @RequestMapping(value = "/rest/comments/{id}",
-            method = RequestMethod.POST,
+            method = RequestMethod.PUT,
             produces = MediaType.APPLICATION_JSON_VALUE)
     @Timed
-    public void update(@PathVariable String id, @Valid @RequestBody CommentDTO commentDTO) {
+    @ApiOperation(value = "Update existing comment")
+    public void update(@PathVariable String id, @Valid @RequestBody CommentDTO commentDTO) throws KalipoRequestException {
         log.debug("REST request to update Comment : {}", commentDTO);
+
+        if (StringUtils.isBlank(id)) {
+            throw new IllegalParameterException();
+        }
+
         Comment comment = CommentDTO.convert(commentDTO);
         comment.setId(id);
-        comment.setAuthorId("d");
+        comment.setAuthorId("d"); // todo SecurityUtils.getCurrentLogin() is null during tests
         comment.setStatus(Comment.Status.APPROVED);
         commentRepository.save(comment);
     }
@@ -71,6 +83,7 @@ public class CommentResource {
             method = RequestMethod.GET,
             produces = MediaType.APPLICATION_JSON_VALUE)
     @Timed
+    @ApiOperation(value = "Get all the comments")
     public List<Comment> getAll() {
 //      todo impl pagination  @QueryParam("offset") int offset, @QueryParam("size") int size
         log.debug("REST request to get all Comments");
@@ -84,8 +97,17 @@ public class CommentResource {
             method = RequestMethod.GET,
             produces = MediaType.APPLICATION_JSON_VALUE)
     @Timed
-    public ResponseEntity<Comment> get(@PathVariable String id) {
+    @ApiOperation(value = "Get the \"id\" comment.")
+    @ApiResponses(value = {
+            @ApiResponse(code = 400, message = "Invalid ID supplied"),
+            @ApiResponse(code = 404, message = "Comment not found")
+    })
+    public ResponseEntity<Comment> get(@PathVariable String id) throws KalipoRequestException {
         log.debug("REST request to get Comment : {}", id);
+        if (StringUtils.isBlank(id)) {
+            throw new IllegalParameterException();
+        }
+
         return Optional.ofNullable(commentRepository.findOne(id))
             .map(comment -> new ResponseEntity<>(
                 comment,
@@ -100,8 +122,15 @@ public class CommentResource {
             method = RequestMethod.DELETE,
             produces = MediaType.APPLICATION_JSON_VALUE)
     @Timed
-    public void delete(@PathVariable String id) {
+    @ResponseStatus(HttpStatus.OK)
+    @ApiOperation(value = "Delete the \"id\" comment")
+    public void delete(@PathVariable String id) throws KalipoRequestException {
         log.debug("REST request to delete Comment : {}", id);
+
+        if (StringUtils.isBlank(id)) {
+            throw new IllegalParameterException();
+        }
+
         commentRepository.delete(id);
     }
 }
